+++ internal/sdkv2provider/schema_cloudflare_access_group.go
@@ -244,7 +244,6 @@ var AccessGroupOptionSchemaElement = &schema.Resource{
 		"external_evaluation": {
 			Type:     schema.TypeList,
 			Optional: true,
-			MaxItems: 1,
 			Elem: &schema.Resource{
 				Schema: map[string]*schema.Schema{
 					"evaluate_url": {